+++ frontend/src/components/courses/CoursesTable.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { AlertTriangle, Loader, X } from 'react-feather';
+import { AlertTriangle, Loader, X, Trash, Edit, ChevronLeft, ChevronRight } from 'react-feather';
 import { useForm } from 'react-hook-form';
 import { Link } from 'react-router-dom';
 
@@ -20,9 +20,10 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
   const { authenticatedUser } = useAuth();
   const [deleteShow, setDeleteShow] = useState<boolean>(false);
   const [isDeleting, setIsDeleting] = useState<boolean>(false);
-  const [selectedCourseId, setSelectedCourseId] = useState<string>();
-  const [error, setError] = useState<string>();
+  const [selectedCourseId, setSelectedCourseId] = useState<string | undefined>();
+  const [error, setError] = useState<string | undefined>();
   const [updateShow, setUpdateShow] = useState<boolean>(false);
+
 
   const {
     register,
@@ -33,45 +34,88 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
   } = useForm<UpdateCourseRequest>();
 
   const handleDelete = async () => {
+    if (!selectedCourseId) {
+      setError("Course ID is required.");
+      return;
+    }
+
     try {
       setIsDeleting(true);
       await courseService.delete(selectedCourseId);
+
       setDeleteShow(false);
-    } catch (error) {
-      setError(error.response.data.message);
+    } catch (error: any) {
+      
+      setError(error?.response?.data?.message || "An unexpected error occurred");
     } finally {
       setIsDeleting(false);
     }
   };
 
   const handleUpdate = async (updateCourseRequest: UpdateCourseRequest) => {
+    if (!selectedCourseId) {
+      setError("Course ID is required.");
+      return;
+    }
+  
+    const formData = new FormData();
+  
+    if (updateCourseRequest.name) formData.append("name", updateCourseRequest.name);
+    if (updateCourseRequest.description) formData.append("description", updateCourseRequest.description);
+    if (updateCourseRequest.type) formData.append("type", updateCourseRequest.type);
+  
+    
+    if (updateCourseRequest.file && updateCourseRequest.file[0]) {
+      formData.append('file', updateCourseRequest.file[0]);  
+    }
+  
     try {
-      await courseService.update(selectedCourseId, updateCourseRequest);
+     
+      await courseService.update(selectedCourseId, formData);
       setUpdateShow(false);
       reset();
-      setError(null);
-    } catch (error) {
-      setError(error.response.data.message);
+      setError(undefined);
+    } catch (error: any) {
+      setError(error?.response?.data?.message || "An unexpected error occurred");
     }
+  };
+  
+  
+
+  // Paginación
+  const [currentPage, setCurrentPage] = useState<number>(1);
+  const itemsPerPage = 3;
+  const totalPages = Math.ceil(data.length / itemsPerPage);
+  const startIndex = (currentPage - 1) * itemsPerPage;
+  const endIndex = currentPage * itemsPerPage;
+  const currentCourses = data.slice(startIndex, endIndex);
+
+  const goToPage = (page: number) => {
+    if (page < 1 || page > totalPages) return;
+    setCurrentPage(page);
   };
 
   return (
     <>
       <div className="table-container">
-        <Table columns={['Name', 'Description', 'Created']}>
+        <Table columns={['Img', 'Name', 'Description', 'Start', 'Type']}>
           {isLoading
             ? null
-            : data.map(({ id, name, description, dateCreated }) => (
+            : currentCourses.map(({ id, name, description, startDate, type, filePath }) => 
+            { console.log(`Curso: ${name}, file:`, filePath); 
+              return(
                 <tr key={id}>
+                  <TableItem>
+                  {filePath && <img src={filePath} alt={name} width="50" />}
+                  </TableItem>
                   <TableItem>
                     <Link to={`/courses/${id}`}>{name}</Link>
                   </TableItem>
                   <TableItem>{description}</TableItem>
-                  <TableItem>
-                    {new Date(dateCreated).toLocaleDateString()}
-                  </TableItem>
+                  <TableItem>{startDate ? new Date(startDate).toLocaleDateString('es-ES') : ''}</TableItem>
+                  <TableItem>{type}</TableItem>
                   <TableItem className="text-right">
-                    {['admin', 'editor'].includes(authenticatedUser.role) ? (
+                      {authenticatedUser?.role && ['admin', 'editor'].includes(authenticatedUser?.role) ? (
                       <button
                         className="text-indigo-600 hover:text-indigo-900 focus:outline-none"
                         onClick={() => {
@@ -83,10 +127,10 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
                           setUpdateShow(true);
                         }}
                       >
-                        Edit
+                        <Edit />
                       </button>
-                    ) : null}
-                    {authenticatedUser.role === 'admin' ? (
+                    ): null}
+                    {authenticatedUser?.role === 'admin' ? (
                       <button
                         className="text-red-600 hover:text-red-900 ml-3 focus:outline-none"
                         onClick={() => {
@@ -94,12 +138,13 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
                           setDeleteShow(true);
                         }}
                       >
-                        Delete
+                      <Trash />
                       </button>
                     ) : null}
                   </TableItem>
                 </tr>
-              ))}
+              )}
+              )}
         </Table>
         {!isLoading && data.length < 1 ? (
           <div className="text-center my-5 text-gray-500">
@@ -107,9 +152,35 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
           </div>
         ) : null}
       </div>
+
+       {/* Paginación */}
+       <div className="flex justify-center gap-2 my-4">
+          <button
+            onClick={() => goToPage(currentPage - 1)}
+            className="p-1 border rounded-lg text-white bg-red-700 hover:bg-red-400 flex items-center gap-2"
+            disabled={currentPage === 1}
+          >
+            <ChevronLeft />
+          </button>
+          <span className="flex items-center">{`Page ${currentPage} of ${totalPages}`}</span>
+          <button
+            onClick={() => goToPage(currentPage + 1)}
+            className="p-1 border rounded-lg text-white bg-red-700 hover:bg-red-400 flex items-center gap-2"
+            disabled={currentPage === totalPages}
+          >
+            <ChevronRight />
+          </button>
+        </div>
+
+        {!isLoading && data.length < 1 ? (
+          <div className="text-center my-5 text-gray-500">
+            <h1>Empty</h1>
+          </div>
+        ) : null}
+      
       {/* Delete Course Modal */}
       <Modal show={deleteShow}>
-        <AlertTriangle size={30} className="text-red-500 mr-5 fixed" />
+        <AlertTriangle width={24} height={24} className="text-red-500 mr-5 fixed" />
         <div className="ml-10">
           <h3 className="mb-2 font-semibold">Delete Course</h3>
           <hr />
@@ -124,7 +195,7 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
           <button
             className="btn"
             onClick={() => {
-              setError(null);
+              setError(undefined);
               setDeleteShow(false);
             }}
             disabled={isDeleting}
@@ -157,11 +228,11 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
             className="ml-auto focus:outline-none"
             onClick={() => {
               setUpdateShow(false);
-              setError(null);
+              setError(undefined);
               reset();
             }}
           >
-            <X size={30} />
+            <X width={24} height={24} />
           </button>
         </div>
         <hr />
@@ -169,22 +240,41 @@ export default function CoursesTable({ data, isLoading }: UsersTableProps) {
         <form
           className="flex flex-col gap-5 mt-5"
           onSubmit={handleSubmit(handleUpdate)}
+          encType="multipart/form-data"
         >
           <input
             type="text"
-            className="input"
+            className="input border"
             placeholder="Name"
             required
             {...register('name')}
           />
           <input
             type="text"
-            className="input"
+            className="input border"
             placeholder="Description"
             required
             disabled={isSubmitting}
             {...register('description')}
           />
+          <select
+            className="input border"
+            disabled={isSubmitting}
+            required
+            {...register('type')}
+          >
+            <option value="">Selecciona una modalidad</option>
+            <option value="Presencial">Presencial</option>
+            <option value="Virtual">Virtual</option>
+          </select>
+
+          <input
+             type="file"
+             className="input border"
+             {...register('file')}
+             disabled={isSubmitting}
+           />
+
           <button className="btn" disabled={isSubmitting}>
             {isSubmitting ? (
               <Loader className="animate-spin mx-auto" />
